+++ src/utils.h
@@ -1,7 +1,11 @@
 /* author: Stéphane Laurent */
+#ifndef UTILSHEADER
+#define UTILSHEADER
+
 
 int cmpfunc(const void*, const void*);
 int cmpfuncdbl(const void*, const void*);
+int cmpedges(const void*, const void*);
 void qsortu(unsigned*, unsigned);
 
 double square(double);
@@ -19,3 +23,7 @@ double dotproduct(double*, double*, unsigned);
 unsigned* uzeros(unsigned);
 
 double squaredDistance(double*, double*, unsigned);
+
+unsigned equalarraysu(unsigned*, unsigned*, unsigned);
+
+#endif